Add schema migration startup mode
