+++ src/server/src/main/docker/entrypoint.sh
@@ -46,6 +46,21 @@ if [ "$1" = 'cassandra-reaper' ]; then
             /etc/cassandra-reaper.yml
 fi
 
+if [ "$1" = 'schema-migration' ]; then
+
+    # get around `/usr/local/bin/configure-persistence.sh: line 65: can't create /etc/cassandra-reaper.yml: Interrupted system call` unknown error
+    touch /etc/cassandra-reaper.yml
+
+    /usr/local/bin/configure-persistence.sh
+    /usr/local/bin/configure-webui-authentication.sh
+    /usr/local/bin/configure-metrics.sh
+    /usr/local/bin/configure-jmx-credentials.sh
+    exec java \
+            ${JAVA_OPTS} \
+            -cp "/usr/local/lib/*" io.cassandrareaper.ReaperApplication schema-migration \
+            /etc/cassandra-reaper.yml
+fi
+
 if [ "$1" = 'register-clusters' ]; then
 
   if [ -z "$2" ]; then
